requirements.txt
aiohttp==3.7.4.post0
discord.py==1.7.3
google==3.0.0
